ci: update the tests to include no-bash and bash non-interactive scenarios
When using multiple `nix-use-shell-action`s in a job, it's common for
bash to disappear from $PATH or for non-interactive bash to be placed
on $PATH (which lacks `compgen` which `action.bash` needs).

Users _can_ work around this by remembering to add `bashInteractive` to
their shells put it'd be better if we just handled this.